Create project.R
On Black Monday, the return on the S&P 500 index was -22.8%. Ouch!
This excercise attempts to answer the question, “what was the conditional probability of a
return this small or smaller on Black Monday?” “Conditional” means given the information
available the previous trading day.
1. Download daily returns on the S&P 500 index from the R package Ecdat using the
following command
data(SP500, package=‘Ecdat’)
You need to install the package first if your computer does not have it installed before.
The S&P 500 returns are in the dataset SP500. The returns are the variable r500,
which spans the period from January 1981 to April 1991.
2. Black Monday is the 1805th return in this dataset. Fit a GARCH(1,1) model to the
last two years of data before Black Monday, assuming 250 trading days per year. What
are the estimates of the parameters of the model?
3. Make a plot of the fitted volatility during the chosen two year period.
4. Calculate the conditional probability of a return less than or equal to -0.228 on Black
Monday.
5. Compute and plot the standardized returns. Also plot the autocorrelations of the
standardized returns and their squares. Do the standardized returns indicate that the
GARCH(1,1) model fits adequately?
6. Are the standardized returns normally distributed? If not, does it have fatter or thinner
tails than a normal distribution? Justify your answers.
7. Fit a GARCH(1,1) model with a t-distributed shock using the following command
fit = garchFit(∼ garch(1,1), data=SP500$r500, cond.dist=‘std’)